pijuice/requirements.txt
balena-sdk==7.1.0